Add jest tooling
Use Jest as the test runner in the "test" script. Jest is a great test runner with very minimal configuration needed to get up and running quickly.

The "test" script allows for other Jest options to be merged into the defaults prior to invoking the test runner.

Update the eslint configuration to allow jest globals in any files that meet the criteria for being a test file. This is so that eslint doesn't error unnecessarily on globals within a test file.